requirements.txt
Fastapi
uvicorn
transformers
torch
pillow
python-multipart
